refactored the code into different files so we can see what we are working with a bit better
